services/api/requirements.txt
fastapi
uvicorn
livekit-api
python-dotenv
python-multipart
